design(inspector) Move Interface panel back to the inspector tab. - Removed `anyKnownElements` and `tabToShow` properties from `ResizableRightPane`. - Readded `SettingsPanel` and carried over `anyKnownElements` to the   `Inspector` component. - Removed `SettingsPanel` from `SettingsPane`.
